Update SDL3 to 15/9 release date, resolve a bunch of warnings all around from MSVC and clang, happier compilers.

+++ src/Systems/CHIP8/Cores/CHIP8_MODERN.cpp
@@ -339,25 +339,25 @@ void CHIP8_MODERN::renderVideoData() {
 	}
 	void CHIP8_MODERN::instruction_8xy5(const s32 X, const s32 Y) noexcept {
 		const bool nborrow{ mRegisterV[X] >= mRegisterV[Y] };
-		mRegisterV[X]   = mRegisterV[X] - mRegisterV[Y];
-		mRegisterV[0xF] = nborrow;
+		mRegisterV[X]   = static_cast<u8>(mRegisterV[X] - mRegisterV[Y]);
+		mRegisterV[0xF] = static_cast<u8>(nborrow);
 	}
 	void CHIP8_MODERN::instruction_8xy7(const s32 X, const s32 Y) noexcept {
 		const bool nborrow{ mRegisterV[Y] >= mRegisterV[X] };
-		mRegisterV[X]   = mRegisterV[Y] - mRegisterV[X];
-		mRegisterV[0xF] = nborrow;
+		mRegisterV[X]   = static_cast<u8>(mRegisterV[Y] - mRegisterV[X]);
+		mRegisterV[0xF] = static_cast<u8>(nborrow);
 	}
 	void CHIP8_MODERN::instruction_8xy6(const s32 X, const s32 Y) noexcept {
 		if (!Quirk.shiftVX) { mRegisterV[X] = mRegisterV[Y]; }
 		const bool lsb{ (mRegisterV[X] & 1) == 1 };
-		mRegisterV[X]   = mRegisterV[X] >> 1;
-		mRegisterV[0xF] = lsb;
+		mRegisterV[X]   = static_cast<u8>(mRegisterV[X] >> 1);
+		mRegisterV[0xF] = static_cast<u8>(lsb);
 	}
 	void CHIP8_MODERN::instruction_8xyE(const s32 X, const s32 Y) noexcept {
 		if (!Quirk.shiftVX) { mRegisterV[X] = mRegisterV[Y]; }
 		const bool msb{ (mRegisterV[X] >> 7) == 1 };
-		mRegisterV[X]   = mRegisterV[X] << 1;
-		mRegisterV[0xF] = msb;
+		mRegisterV[X]   = static_cast<u8>(mRegisterV[X] << 1);
+		mRegisterV[0xF] = static_cast<u8>(msb);
 	}
 
 	#pragma endregion
