+++ src/Systems/CHIP8/Cores/XOCHIP.cpp
@@ -297,10 +297,11 @@ void XOCHIP::renderVideoData() {
 		textureBuffer.end(),
 		[&](u8& pixel) noexcept {
 			const auto idx{ &pixel - textureBuffer.data() };
-			pixel = mDisplayBuffer[3].at_raw(idx) << 3
-				  | mDisplayBuffer[2].at_raw(idx) << 2
-				  | mDisplayBuffer[1].at_raw(idx) << 1
-				  | mDisplayBuffer[0].at_raw(idx);
+			pixel = static_cast<u8>(
+				mDisplayBuffer[3].at_raw(idx) << 3 |
+				mDisplayBuffer[2].at_raw(idx) << 2 |
+				mDisplayBuffer[1].at_raw(idx) << 1 |
+				mDisplayBuffer[0].at_raw(idx));
 		}
 	);
 
@@ -549,25 +550,25 @@ void XOCHIP::scrollDisplayRT() {
 	}
 	void XOCHIP::instruction_8xy5(const s32 X, const s32 Y) noexcept {
 		const bool nborrow{ mRegisterV[X] >= mRegisterV[Y] };
-		mRegisterV[X]   = mRegisterV[X] - mRegisterV[Y];
-		mRegisterV[0xF] = nborrow;
+		mRegisterV[X]   = static_cast<u8>(mRegisterV[X] - mRegisterV[Y]);
+		mRegisterV[0xF] = static_cast<u8>(nborrow);
 	}
 	void XOCHIP::instruction_8xy7(const s32 X, const s32 Y) noexcept {
 		const bool nborrow{ mRegisterV[Y] >= mRegisterV[X] };
-		mRegisterV[X]   = mRegisterV[Y] - mRegisterV[X];
-		mRegisterV[0xF] = nborrow;
+		mRegisterV[X]   = static_cast<u8>(mRegisterV[Y] - mRegisterV[X]);
+		mRegisterV[0xF] = static_cast<u8>(nborrow);
 	}
 	void XOCHIP::instruction_8xy6(const s32 X, const s32 Y) noexcept {
 		if (!Quirk.shiftVX) { mRegisterV[X] = mRegisterV[Y]; }
 		const bool lsb{ (mRegisterV[X] & 1) == 1 };
-		mRegisterV[X]   = mRegisterV[X] >> 1;
-		mRegisterV[0xF] = lsb;
+		mRegisterV[X]   = static_cast<u8>(mRegisterV[X] >> 1);
+		mRegisterV[0xF] = static_cast<u8>(lsb);
 	}
 	void XOCHIP::instruction_8xyE(const s32 X, const s32 Y) noexcept {
 		if (!Quirk.shiftVX) { mRegisterV[X] = mRegisterV[Y]; }
 		const bool msb{ (mRegisterV[X] >> 7) == 1 };
-		mRegisterV[X]   = mRegisterV[X] << 1;
-		mRegisterV[0xF] = msb;
+		mRegisterV[X]   = static_cast<u8>(mRegisterV[X] << 1);
+		mRegisterV[0xF] = static_cast<u8>(msb);
 	}
 
 	#pragma endregion
